+++ app.rb
@@ -14,27 +14,49 @@ class App
     @rentals = []
   end
 
-  def run
-    puts 'Welcome to the Console Library App!'
-    loop do
-      display_options
-      choice = gets.chomp.to_i
-      handle_choice(choice)
-      break if choice == 7
+  # def run
+  #   puts 'Welcome to the Console Library App!'
+  #   loop do
+  #     display_options
+  #     choice = gets.chomp.to_i
+  #     handle_choice(choice)
+  #     break if choice == 7
+  #   end
+  #   puts 'Thank you for using the Console Library App. Goodbye!'
+  # end
+
+  # def handle_choice(choice)
+  #   case choice
+  #   when 1 then list_all_books
+  #   when 2 then list_all_people
+  #   when 3 then create_person
+  #   when 4 then create_book
+  #   when 5 then create_rental
+  #   when 6 then list_rentals_for_person
+  #   else
+  #     puts 'Invalid option. Please try again.'
+  #   end
+  # end
+
+  def list_all_books
+    if @books.empty?
+      puts 'No books available.'
+    else
+      puts 'Listing all books:'
+      @books.each_with_index do |book, index|
+        puts "#{index}. #{book.title} by #{book.author}"
+      end
     end
-    puts 'Thank you for using the Console Library App. Goodbye!'
   end
 
-  def handle_choice(choice)
-    case choice
-    when 1 then list_all_books
-    when 2 then list_all_people
-    when 3 then create_person
-    when 4 then create_book
-    when 5 then create_rental
-    when 6 then list_rentals_for_person
+  def list_all_people
+    if @people.empty?
+      puts 'No people available.'
     else
-      puts 'Invalid option. Please try again.'
+      puts 'Listing all people:'
+      @people.each_with_index do |person, index|
+        puts "#{index}. #{person.name} - (ID: #{person.id})"
+      end
     end
   end
